Updated based on code review
Created ComposedTaskRunnerVisitor in the DSL module in a new visitor package.
Rebased
