+++ spring-cloud-dataflow-server-core/src/main/java/org/springframework/cloud/dataflow/server/service/impl/DefaultTaskExecutionService.java
@@ -42,11 +42,9 @@ import org.springframework.cloud.dataflow.core.TaskDefinition;
 import org.springframework.cloud.dataflow.core.TaskDeployment;
 import org.springframework.cloud.dataflow.core.TaskManifest;
 import org.springframework.cloud.dataflow.core.TaskPlatformFactory;
-import org.springframework.cloud.dataflow.core.dsl.SplitNode;
-import org.springframework.cloud.dataflow.core.dsl.TaskAppNode;
 import org.springframework.cloud.dataflow.core.dsl.TaskNode;
 import org.springframework.cloud.dataflow.core.dsl.TaskParser;
-import org.springframework.cloud.dataflow.core.dsl.TaskVisitor;
+import org.springframework.cloud.dataflow.core.dsl.visitor.ComposedTaskRunnerVisitor;
 import org.springframework.cloud.dataflow.rest.util.ArgumentSanitizer;
 import org.springframework.cloud.dataflow.rest.util.DeploymentPropertiesUtils;
 import org.springframework.cloud.dataflow.server.job.LauncherRepository;
@@ -713,47 +711,13 @@ public class DefaultTaskExecutionService implements TaskExecutionService {
 	private void isCTRSplitValidForCurrentCTR(TaskLauncher taskLauncher, TaskDefinition taskDefinition) {
 		TaskParser taskParser = new TaskParser("composed-task-runner", taskDefinition.getProperties().get("graph"), true, true);
 		TaskNode taskNode = taskParser.parse();
-		ComposedRunnerVisitor composedRunnerVisitor = new ComposedRunnerVisitor();
+		ComposedTaskRunnerVisitor composedRunnerVisitor = new ComposedTaskRunnerVisitor();
 		taskNode.accept(composedRunnerVisitor);
-		if(composedRunnerVisitor.highCount > taskLauncher.getMaximumConcurrentTasks()) {
+		if(composedRunnerVisitor.getHighCount() > taskLauncher.getMaximumConcurrentTasks()) {
 			throw new IllegalArgumentException(String.format("One or more of the " +
 					"splits in the composed task contains a task count that exceeds " +
 					"the maximumConcurrentTasks count of %s",
 					taskLauncher.getMaximumConcurrentTasks()));
 		}
 	}
-
-
-	private class ComposedRunnerVisitor extends TaskVisitor {
-
-		private int highCount;
-		private int currentCount;
-
-		/**
-		 * Reset current counter for a new split
-		 */
-		public void visit(SplitNode split) {
-			logger.debug("Visit Split:  " + split);
-			currentCount = 0;
-
-		}
-
-		/**
-		 * At end of split determine if the number of task apps in the split is greater than other splits.
-		 */
-		public void postVisit(SplitNode split) {
-			if(currentCount > highCount) {
-				highCount = currentCount;
-			}
-		}
-
-		/**
-		 * Count the number of tasks in a split.
-		 */
-		public void visit(TaskAppNode taskApp) {
-			currentCount++;
-		}
-
-	}
-
 }
